Add ability to recognize \N as a Token::CharClass::Simple, introduced in Perl 5.11.
Correct the Token::Literal logic to recognize a Unicode character by
name ( \N{name} ) only if \N is followed by a left curly and a letter,
as documented in perl5110delta (previous logic was a curly and a \w, \s
or ':').

Document in PPIx::Regexp::Tokenizer the new ordering constraint
(Token::CharClass::Simple has to come after Token::Literal).

Add a few tests to t/unit.t to make sure \N is recognized, and returns
the correct value for perl_version_introduced().


git-svn-id: file:///Users/tom/Code/svn/repos/PPIx-Regexp/trunk@1503 f6fd3633-1913-47c2-bad4-372d75c948ec