RT 56864 PPIx::Regexp::Lexer fails in Perl::Critic under Perl 5.13.0
This is one of those where I fixed it but I'm not sure why what I did
worked. The problem was that the tokenizer's find_regexp() method was
returning $+[0], and under some yet-to-be-elucidated circumstances (I
never duplicated the error outside Perl::Critic) the caller did not see
the same value the subroutine was trying to return. The fix was to stage
the data in a lexical variable, and return that.

The symptoms were caused by the tokenizer sometimes grabbing more
characters than it was supposed to. When it grabbed a string containing
a left parenthesis and called it all white space, the lexer became very,
very confused.

Also included in this update is support for setting the tokenizer trace
attribute from environment variable PPIX_REGEXP_TOKENIZER_TRACE, and
some checks in the lexer that call Carp::confess (hopefully with useful
information) at the point the code was blowing up.


git-svn-id: file:///Users/tom/Code/svn/repos/PPIx-Regexp/trunk@1749 f6fd3633-1913-47c2-bad4-372d75c948ec