Fix inconsistency in perl_version_introduced() between PPIx::Regexp::Token::Modifier and PPIx::Token::GroupType::Modifier by having the latter delegate to the former and then modify the results as necessary.
Add support for the Perl 5.13.6 construction (?^...) and the new
modifiers d, l, and u, which are only valid inside (?...). This includes
not only the modifier-juggling but the match_semantics() method to
return the modifier currently in effect. The actual implementation may
undergo further changes, but what's committed is perhaps not completely
shabby.

As part of the above, the guts of PPIx::Regexp::Tokenizer's
modifier_modify() logic got moved to PPIx::Regexp::Token::Modifier's
__PPIX_TOKENIZER__modifier_modify method, so that all the match
semantics modifier juggling could be in one place.

Modified PPIx::Regexp::Test (in inc/) to display the actual arguments in
the true(), false(), and value() subroutines. The latter also displays
the expected result.


git-svn-id: file:///Users/tom/Code/svn/repos/PPIx-Regexp/trunk@2066 f6fd3633-1913-47c2-bad4-372d75c948ec