Modify PPIx::Regexp::Token::Modifier to carry aggregated modifiers (like 'd', 'l', 'u') in aggregated form internally. These are de-aggregated for the public interface.
This means that asserted('d') returns true if in fact the 'd' modifier
is in effect, even though there is no 'd' key in the internal modifiers
hash. On the other hand, asserted('match_semantics') will return false,
even though one of the 'd', 'l', or 'u' modifiers is in efect and these
are (currently)  stored under key 'match_semantics'.

Similarly, the hash (or hash reference) returned by modifiers() will
never contain the 'match_semantics' key. If that key is present it will
contain a key representing the modifier that is the value of the
'match_semantics' key.

But the modifiers hash maintained by the tokenizer during the token scan
_will_ have a 'match_semantics' key if one of 'd', 'l', or 'u' is in
effect, since this is internal to PPIx::Regexp.


git-svn-id: file:///Users/tom/Code/svn/repos/PPIx-Regexp/trunk@2067 f6fd3633-1913-47c2-bad4-372d75c948ec