Make t/version.t into a dual-purposed script. Its primary purpose is to test that the perl_version_introduced() and perl_version_removed() methods return the correct value. But it has also been given the ability, if $REPORT is true, to report on these.
I have misgivings about this, because of all the stuff that needed to be
added for the secondary purpose that is not needed for the primary
purpose. The result has kind of a tail-wagging-the-dog feel. At this
point I have no other ideas, though, other than to maintain a database
somehow of all the information, and then generate the test and the
report from the database. And I'm not yet ready to take that step.


git-svn-id: file:///Users/tom/Code/svn/repos/PPIx-Regexp/trunk@2644 f6fd3633-1913-47c2-bad4-372d75c948ec