Add support for default modifiers, to (eventually) deal with Perl-Critic RT 72151. This support consists of:
New argument default_modifiers to the new() methods of PPIx::Regexp,
PPIx::Regexp::Tokenizer, and PPIx::Regexp::Dumper.

PPIx::Regexp::Dumper support for displaying the default_modifiers
argument.

New public method modifier_asserted on PPIx::Regexp. This returns
whether the modifier is actually in effect. The modifier() method
continues to return the actual expression modifier, for round-trip
purposes.

Tests for all this.

Support for all this in eg/predump.

Various package-private methods to support encapsulation while passing
the necessary information from where it is generated to where it needs
to be.


git-svn-id: file:///Users/tom/Code/svn/repos/PPIx-Regexp/trunk@2645 f6fd3633-1913-47c2-bad4-372d75c948ec