Fix RT#82140: incorrect parsing of \? - Alexandr Ciornii
The problem was that I tried to do too much in the regexp that matched
the group type. The solution was to have PPIx::Regexp::Tokenizer make
the start delimiter available, and then have
PPIx::Regexp::Token::GroupType::BranchReset make use of this information
to choose a regexp appropriate to the delimiter.

There may be more of these around.


git-svn-id: file:///Users/tom/Code/svn/repos/PPIx-Regexp/trunk@3682 f6fd3633-1913-47c2-bad4-372d75c948ec