Properly handle absolute backrefs like \g<number>
The code was converting these to literals if the number was greater than
the actual number of capture groups. This is wrong, since Perl fails to
compile them. So they are being converted to unknown tokens, and an
error declared. To be consistent I should do relative backreferences and
named backreferences as well. Named backreferences are easy. Relative
ones are less so.