RT #91798: Incorrect handling of white space
There were two problems here.

The first was that I misread perl5179delta to say that the additional
characters parsed as white space under /x were implemented in that
release. This was false; what that release of Perl did was to warn when
the additional characters were found, but still treat them as literals.

The second was that \s was matching more than I thought it was, and so
too many characters were considered white space.

This change was made in commit ad9b35bf, and released in version 0.033
on February 22 2013.

Thanks to Nobuo Kumagai for finding and reporting this.