Track number of times modifier asserted.
This affects the tokenizer's modifier() method, and the
::Token::Modifier asserted() method. In both cases, if the modifier
being queried is not one of the match semantics modifier (or, in theory,
any modifier in an aggregated group, of which so far this is the only
one), if you follow the modifier by an asterisk you get the number of
times that modifier is asserted. For example,

 $tokenizer->modifier( 'e*' )

returns the number of /e modifiers found (i.e. 1 for /e, 2 for /ee, and
so on), whereas

 $tokenizer->modifier( 'e' )

retains the old behavior of returning true for /e and false for /ee.

This was triggered by two considerations:

1) The plan (announced with Perl 5.21.5) of having /xx allow white space
in character classes

2) The observation of the undocumented (as far as I know) fact that
s///eee in fact evals the replacement string twice, though /ee only
evals it once.