RT 107331 Erroneous parse due to trailing cruft
The issue is that PPIx::Regexp assumes you have handed it a valid
regular expression of some sort. The subject ticket was actually
retracted by its submitter after he observed the problem was due to a
trailing single quote. PPIx::Regexp blithely interpreted this as the
closing delimiter and never noticed the mismatch.

Thanks to Klaus Rindfrey for catching this.

The fix was to have the lexer check the delimiters. If they do not match
the trailing one is reblessed to PPIx::Regexp::Token::Unknown and the
error count is incremented. The dumper object was also enhanced to look
for such tokens as delimiters, and display the error text if any.

An alternate implementation was to use the PPI parser engine to validate
the input. This has the advantage that I could simply manufacture a
token to hold the trailing cruft and the error would be more obvious in
the dump. But this is also riskier since I do not know that the PPI
parser behaves the same as my tokenizer, and because any problems become
PPI bugs and mean that I have to hope PPI fixes them.