Fixed but in parenthesis group type matching
Basically these are recognized by regular expressions, but the regular
expression varies based on the delimiter character used (e.g. m/(?<x)/
and m<(?\<x)> are equivalent, but need different matching regexes).

The problem was that the regexes were generated from a list of valid
group type strings, but some necessary regexes were not being generated.
The solution was to move the computation of the list of characters to
escape outside the loop that generates the regexes.