Consolidate tokenizer code for trailing tokens
The issue here was that in the previous implementation we needed to
recognize the modifiers in two places: in the 'init' phase because we
need them to correctly tokenize, and in the 'finish' phase to actually
append them to the token stream.

This modification manufactures the modifier token, as well as the
trailing white space and other cruft, in the 'init' phase and then
stashes them. When the 'finish' phase is executed, the stashed tokens
are appended to the token stream and the scan set to the end of the
string.