Correct mis-parse of s/// with leading space
The PPIx::Regexp::Lexer lex() method incorrectly used the first token's
content to detect a substitution expression. But this failed when the
first token was white space, which is permitted. Instead we pick up the
first PPIx::Regexp::Token::Structure, and cunsult that if it was found.
If it was not, we can not have a substitution, and in fact probably have
the case where the expression parsed in its entirety into the unknown
token.