Correct tokenizer where entire expression unknown
In the cases where the tokenizer can not even get off the ground, the
tokenizer makes the whole expression into a single
PPIx::Regexp::Token::Unknown. The tokenizer was then put into 'kaput'
state (meaning no more tokens were expected), but the cursor was not
moved to the end of the content, so the tokenizer tried to process them
again, resulting in an error because the 'kaput' state does not have a
handler.

The fix was to centralize the code that changes the state, and have that
code manually move the cursor to the end of the content when the state
goes to 'kaput'.

This was picked up when I decided it would be interesting to fuzz test.