Correct code faulure when parsing all white space
The problem was that when the tokenizer was instantiated the cursor
limit was moved to the left of trailing white space. This got fixed if
the string being tokenized contained anything other than white space. If
it did not, we ended up dropping into the finalization code immediately,
and ultimately dying.