Document policy for incoompatible parses
The problem is that sometimes Perl changes in such a way that the proper
parse for a regular expression changes. In this case PPIx::Regexp
produces the more modern parse. Documented this, along with known
examples: the non-interpolation of $(, $), $|, @+, and @-. Subscripted
versions of the last two do interpolate, and are so parsed.