Rationalize reblessing machinery
The last commit left two methods with the same name but different
signatures, both called by the lexer but under different circumstances.
These have been made consistent as __PPIX_ELEM__rebless(), which takes
a class name and optional named arguments and returns the number of
errors detected. The PPIx::Regexp::Element version either clears or sets
the {error} attribute depending on the presence and value of the 'error'
argument, and returns 0 or 1 respectively.

The PPIx::Regexp::Token::Backreference incarnation of this (which
antedates the current round of revisions) gained the class name as the
first post-invocant argument. If this is defined, it simply calls the
super method, returning its result. If this is undef, we get the old
behavior, in which it computes the class name.