Reorder arguments of __PPIX_ELEM__rebless()
Previously the invocant was the object to rebless and the first argument
was the name of the class to bless into. This change makes the invocant
the class to bless into and the first argument the object or hash
reference to bless, which potentially means I can use it as an
instantiator.