Add explain() and supporting methods.
The explain() method returns a brief explanation of what the element
does.

main_structure() returns the main structure that contains the invocant,
or nothing if the invocant is not contained in a main structure. The
main structures are the PPIx::Regexp::Structure::Main elements, which at
the moment are PPIx::Regexp::Structure::Regexp and
PPIx::Regexp::Structure::Replacement. Main structures do not contain
themselves.

in_regex_set() returns a true value if the invocant is contained in a
regexp set (a.k.a. extended bracketed character class) at any level of
nesting. Extended bracketed character classes are considered to contain
themselves.