Always provide tokenizer in t/version.t
This test manufactures tokens ad-hoc, rather than by tokenizing. This
modification makes sure there is always a tokenizer available for
__PPIX_TOKEN__post_make(), and eliminates all code in that method that
checks for the existence of a tokenizer.

Because the dummy tokenizer is not properly initialized, it was
necessary to run any token-matching logic by hand in t/version.t to make
sure any required captures are populated.