Recognize postfix dereference if desired
This is specified by the Boolean 'postderef' argument to the tokenizer
and anyone who uses it. This affects how interpolations are parsed, and
the version returned by interpolations and code, since postfix
dereference was introduced in 5.19.5.

The default is false, but will become true if postfix dereferences
become mainstream. It (the default) is the value of
$PPIx::Regexp::Tokenizer::DEFAULT_POSTDEREF.