Provide mock tokenizer to support instantiation
The real incentive for this is that t/basic.t is kind of a basic unit
test, but the previous commit (ditching __PPIX_TOKEN__post_make() in
favor of doing all the requisit logic in __new()) introduced a
chicken-and-egg problem if I wanted to actually instantiate in that
test. So actual instantiations of classes that needed data from the
tokenizer were commented out.

The mock tokenizer simply returns data given it when it is instantiated,
and lets me add the instantiation tests back (except for one that was
eliminated because it instantiated an abstract class). This might also
be a better way to handle t/version.t than to instantiate a real
tokenizer.