Deprecate tokenizer method prior()
This is deprecated in favor of prior_significant_token(). In six months
after the next release, it will warn on the first use. Six months after
that it will warn on each use. Six months after this it will be fatal.

I am not sure that I need to do this, since prior() (and
prior_significant_token(), for that matter) are documented as not being
part of the public interface.