Parse '\N{}' as no-op.
The previous parse was a character class ('\N') followed by two literals
('{' and '}'). But perl5238delta said that it had been ignored up to
that time. Starting with 5.23.8 it is an error if 'use re strict' is in
effect.

The easiest implementation was to accept it as a literal, but it really
isn't. The next easiest was as white space, but it isn't that either.
That left making it a whole new class, which got inserted in the
hierarchy between PPIx::Regexp::Token and
PPIx::Regexp::Token::Whitespace.