Add option 'strict', like 'use re "strict"'
This has its fingers in everything, since the signature of
__PPIX_LEXER__finalize() had to include the lexer object so the elements
could determine if strict() was in effect. It also meant another new
class, PPIx::Regexp::Node::Unknown, since I needed to rebless
PPIx::Regexp::Node::Range into something if the end points were not
portable and compatible.

In the presence of strict(), I opted to set perl_version_introduced to
the version of Perl where the construct became an error.