Support deprecating left curlys after left parens
These were originally going to be OK, but in 5.27.8 that changed. The
logic is in place, and tested by such a horrible hack that I made it an
author test, so when they finally go away I should be able to set the
correct value in the relevant manifest constant and everything will Just
Work.