Explain grouping structure as 'Grouping'.
The previous explanation for a bare 'PPIx::Regexp::Structure' was
'Grouping or capture,' but if it captureed it would be a
'PPIx::Regexp::Structure::Capture'. The 'Grouping or capture'
explanation is still given for the tokenizer output, because with just a
stream of tokens we do not know whether the parens capture or not.