Install @CARP_NOT everywhere
so that warnings and exceptions generated in the bowels of the system
appear to come from the point where the system is entered.

An author test was added to make sure this policy continues.

The incentive was that the deprecation message for the 'parse' argument
comes from PPIx::Regexp::Util, which is at least three modules in from
the point where the 'parse' argument was specified. I may have overdone
it because I installed @CARP_NOT even in essentially-empty classes
having no descendants, but that way I don't leave a trap for myself if
their situation changes.