Support proper version for qr'\N{name}'.
Until 5.29.10 this construction failed to parse because it did not
interpolate. But PPIx::Regexp blithely ignored this detail. As of
5.29.10, something like m'\N{LATIN CAPITAL LETTER L}' matches
identically to m'L'. So I implemented that.

Before 5.29.10 the proper implementation was to declare it an error,
since as far back as I have looked (5.10) Perl called it an error,
rather than parsing it as individual characters.