Use 'use open ...' to set test handles to utf-8.
This must be done before Test::More is loaded, and since it was
conditional is a require() in a BEGIN block, and using strings rather
than barewords because of Perl's propensity to parse bareword open as
CORE::open().

The previous implementation waited until Test::More had been loaded, and
then called Test::More->builder() to get the Test::Builder singleton and
manipulated all its handles.